+++ app/resources/views/pages/dominion/town-crier.blade.php
@@ -126,11 +126,11 @@
                 </div>
                 <div class="box-body">
                     @if ($fromOpCenter)
-                        <p>All the news for the target's realm will be shown here. The town crier presents news for a 2 day period.</p>
+                        <p>All the news for the target's realm will be shown here.</p>
                     @else
-                        <p>All the news for your realm can be seen here. The town crier gives you news for a 2 day period.</p>
+                        <p>All the news for your realm can be seen here.</p>
                     @endif
-                    <p>You will see only military operations, as well as death messages{{-- and important messages regarding Wonders of the World--}}. Magical and Spy attacks are not known to the Town Crier, and you will have to inquire in the council as to those types of attacks.</p>
+                    <p>You will see only military operations{{-- as well as death messages and important messages regarding Wonders of the World--}}. Magical and Spy attacks are not known to the Town Crier. You will have to inquire in the council as to those types of attacks.</p>
                     @if ($selectedDominion->round->start_date <= now())
                     <p>
                         <label for="realm-select">Show Town Crier for:</label>